Break out InventoryHighlightPrefabInterceptWidgetExt
